Turn the roomservice on/off the easy way
Up until now the only way to really turn off the roomservice was to delete
the following lines from the envsetup script

https://github.com/DirtyUnicorns/android_build/blob/o8x/envsetup.sh#L668-L672

While this was great it caused a problem that if you wanted to say commit something
on to the build repo, you had to add back those lines. Then if you wanted to turn
off the roomservice again, you guessed it...you had to remove those lines. Really it
seem like the cave man way of doing things.

"Who the hell does this anyways?"

A lot of people when they're working on bringing up a device or rebasing. If you're
bringing up a device, is usually not in the main org and so the dependencies will fail
due to the roomservice not finding those repos. Sure we can adjust as we have been
doing
for years but why when you can just type in one command and BOOM! The source is
maintained
untouched and you can toggle it on and off to fit your needs!

- To turn off the roomservice just type in NO_ROOMSERVICE=true

- To turn if back on just type in NO_ROOMSERVICE=false

The roomservice is on by default as it should!

Change-Id: I478d03fb23c0188fe8ec37b124700d0e317d0705